feat: reorganize patterns and usecases in their own files
